# docs/swagger.yaml
definitions:
...
  types.PodListResponse:
    properties:
      age:
        type: string
      clusterName:
        type: string
      configName:
        type: string
      cpu:
        type: string
      hasUpdated:
        type: boolean
      lastRestartAt:
        type: string
      lastRestartReason:
        type: string
      memory:
        type: string
      name:
        type: string
      namespace:
        type: string
      node:
        type: string
      podIP:
        type: string
      qos:
        type: string
      ready:
        type: string
      restarts:
        type: string
      status:
        type: string
      uid:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/namespaces:
    get:
      consumes:
      - text/event-stream
      description: Streams Namespaces data in real-time using Server-Sent Events.
        Provides live updates of namespace status.
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name for multi-cluster setups
        in: query
        name: cluster
        type: string
      produces:
      - text/event-stream
      - application/json
      responses:
        "200":
          description: Streaming Namespaces data
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Namespaces (SSE)
      tags:
      - Cluster
  /api/v1/namespaces/{name}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific namespace
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name for multi-cluster setups
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: Namespace details
          schema:
            type: object
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Namespace by name
      tags:
      - Cluster
  /api/v1/namespaces/{name}/events:
    get:
      consumes:
      - application/json
      description: Retrieves events related to a specific namespace
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name for multi-cluster setups
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: Namespace events
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Namespace events
      tags:
      - Cluster
  /api/v1/namespaces/{name}/pods:
    get:
      consumes:
      - application/json
      - text/event-stream
      description: Retrieves all pods in a specific namespace with real-time updates
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name for multi-cluster setups
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: Namespace pods
          schema:
            items:
              $ref: '#/definitions/types.PodListResponse'
            type: array
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Namespace pods
      tags:
      - Cluster
  /api/v1/namespaces/{name}/yaml:
    get:
      consumes:
      - application/json
      description: Retrieves the YAML representation of a specific namespace
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name for multi-cluster setups
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: path
        name: name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Namespace YAML representation
          schema:
            type: string
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Namespace YAML
      tags:
      - Cluster
...
swagger: "2.0"